Update to bsnes v107r6 beta release.
byuu says:

  - converted (u)int(8,16,32,64) from Natural/Integer<T> to
    (u)int(8,16,32,64)_t types
  - SFC: mostly rewritten WDC65816 CPU core
  - removed 487KiB of code! (unused CPU cores from other higan cores)

+++ bsnes/processor/gsu/gsu.hpp
@@ -20,8 +20,8 @@ struct GSU {
   virtual auto writeRAMBuffer(uint16 addr, uint8 data) -> void = 0;
   virtual auto flushCache() -> void = 0;
 
-  virtual auto read(uint24 addr, uint8 data = 0x00) -> uint8 = 0;
-  virtual auto write(uint24 addr, uint8 data) -> void = 0;
+  virtual auto read(uint addr, uint8 data = 0x00) -> uint8 = 0;
+  virtual auto write(uint addr, uint8 data) -> void = 0;
 
   //gsu.cpp
   auto power() -> void;
