+++ bsnes/sfc/coprocessor/cx4/cx4.hpp
@@ -1,8 +1,8 @@
 struct Cx4 {
   auto power() -> void;
 
-  auto read(uint24 addr, uint8 data = 0) -> uint8;
-  auto write(uint24 addr, uint8 data) -> void;
+  auto read(uint addr, uint8 data = 0) -> uint8;
+  auto write(uint addr, uint8 data) -> void;
 
   auto serialize(serializer&) -> void;
