+++ bsnes/sfc/coprocessor/event/event.hpp
@@ -20,11 +20,11 @@ struct Event : Thread {
   auto unload() -> void;
   auto power() -> void;
 
-  auto mcuRead(uint24 addr, uint8) -> uint8;
-  auto mcuWrite(uint24 addr, uint8) -> void;
+  auto mcuRead(uint addr, uint8) -> uint8;
+  auto mcuWrite(uint addr, uint8) -> void;
 
-  auto read(uint24 addr, uint8 data) -> uint8;
-  auto write(uint24 addr, uint8 data) -> void;
+  auto read(uint addr, uint8 data) -> uint8;
+  auto write(uint addr, uint8 data) -> void;
 
   //serialization.cpp
   auto serialize(serializer&) -> void;
