+++ bsnes/sfc/coprocessor/hitachidsp/hitachidsp.hpp
@@ -10,32 +10,32 @@ struct HitachiDSP : Processor::HG51B, Thread {
   auto unload() -> void;
   auto power() -> void;
 
-  auto isROM(uint24 address) -> bool override;
-  auto isRAM(uint24 address) -> bool override;
+  auto isROM(uint address) -> bool override;
+  auto isRAM(uint address) -> bool override;
 
   //HG51B read/write
-  auto read(uint24 address) -> uint8 override;
-  auto write(uint24 address, uint8 data) -> void override;
+  auto read(uint address) -> uint8 override;
+  auto write(uint address, uint8 data) -> void override;
 
   //CPU ROM read/write
-  auto addressROM(uint24 address) const -> maybe<uint24>;
-  auto readROM(uint24 address, uint8 data = 0) -> uint8;
-  auto writeROM(uint24 address, uint8 data) -> void;
+  auto addressROM(uint address) const -> maybe<uint>;
+  auto readROM(uint address, uint8 data = 0) -> uint8;
+  auto writeROM(uint address, uint8 data) -> void;
 
   //CPU RAM read/write
-  auto addressRAM(uint24 address) const -> maybe<uint24>;
-  auto readRAM(uint24 address, uint8 data = 0) -> uint8;
-  auto writeRAM(uint24 address, uint8 data) -> void;
+  auto addressRAM(uint address) const -> maybe<uint>;
+  auto readRAM(uint address, uint8 data = 0) -> uint8;
+  auto writeRAM(uint address, uint8 data) -> void;
 
   //HG51B data RAM read/write
-  auto addressDRAM(uint24 address) const -> maybe<uint24>;
-  auto readDRAM(uint24 address, uint8 data = 0) -> uint8;
-  auto writeDRAM(uint24 address, uint8 data) -> void;
+  auto addressDRAM(uint address) const -> maybe<uint>;
+  auto readDRAM(uint address, uint8 data = 0) -> uint8;
+  auto writeDRAM(uint address, uint8 data) -> void;
 
   //CPU IO read/write
-  auto addressIO(uint24 address) const -> maybe<uint24>;
-  auto readIO(uint24 address, uint8 data = 0) -> uint8;
-  auto writeIO(uint24 address, uint8 data) -> void;
+  auto addressIO(uint address) const -> maybe<uint>;
+  auto readIO(uint address, uint8 data = 0) -> uint8;
+  auto writeIO(uint address, uint8 data) -> void;
 
   auto firmware() const -> vector<uint8>;
   auto serialize(serializer&) -> void;
