+++ bsnes/sfc/coprocessor/mcc/mcc.hpp
@@ -10,17 +10,17 @@ struct MCC {
   auto power() -> void;
   auto commit() -> void;
 
-  auto read(uint24 address, uint8 data) -> uint8;
-  auto write(uint24 address, uint8 data) -> void;
+  auto read(uint address, uint8 data) -> uint8;
+  auto write(uint address, uint8 data) -> void;
 
-  auto mcuRead(uint24 address, uint8 data) -> uint8;
-  auto mcuWrite(uint24 address, uint8 data) -> void;
+  auto mcuRead(uint address, uint8 data) -> uint8;
+  auto mcuWrite(uint address, uint8 data) -> void;
 
-  auto mcuAccess(bool mode, uint24 address, uint8 data) -> uint8;
-  auto romAccess(bool mode, uint24 address, uint8 data) -> uint8;
-  auto psramAccess(bool mode, uint24 address, uint8 data) -> uint8;
-  auto exAccess(bool mode, uint24 address, uint8 data) -> uint8;
-  auto bsAccess(bool mode, uint24 address, uint8 data) -> uint8;
+  auto mcuAccess(bool mode, uint address, uint8 data) -> uint8;
+  auto romAccess(bool mode, uint address, uint8 data) -> uint8;
+  auto psramAccess(bool mode, uint address, uint8 data) -> uint8;
+  auto exAccess(bool mode, uint address, uint8 data) -> uint8;
+  auto bsAccess(bool mode, uint address, uint8 data) -> uint8;
 
   //serialization.cpp
   auto serialize(serializer&) -> void;
