+++ bsnes/sfc/coprocessor/sa1/iram.cpp
@@ -5,32 +5,32 @@ auto SA1::IRAM::conflict() const -> bool {
   return false;
 }
 
-auto SA1::IRAM::read(uint24 address, uint8 data) -> uint8 {
+auto SA1::IRAM::read(uint address, uint8 data) -> uint8 {
   if(!size()) return data;
   address = bus.mirror(address, size());
   return WritableMemory::read(address, data);
 }
 
-auto SA1::IRAM::write(uint24 address, uint8 data) -> void {
+auto SA1::IRAM::write(uint address, uint8 data) -> void {
   if(!size()) return;
   address = bus.mirror(address, size());
   return WritableMemory::write(address, data);
 }
 
-auto SA1::IRAM::readCPU(uint24 address, uint8 data) -> uint8 {
+auto SA1::IRAM::readCPU(uint address, uint8 data) -> uint8 {
   cpu.synchronize(sa1);
   return read(address, data);
 }
 
-auto SA1::IRAM::writeCPU(uint24 address, uint8 data) -> void {
+auto SA1::IRAM::writeCPU(uint address, uint8 data) -> void {
   cpu.synchronize(sa1);
   return write(address, data);
 }
 
-auto SA1::IRAM::readSA1(uint24 address, uint8 data) -> uint8 {
+auto SA1::IRAM::readSA1(uint address, uint8 data) -> uint8 {
   return read(address, data);
 }
 
-auto SA1::IRAM::writeSA1(uint24 address, uint8 data) -> void {
+auto SA1::IRAM::writeSA1(uint address, uint8 data) -> void {
   return write(address, data);
 }
