+++ bsnes/sfc/coprocessor/spc7110/spc7110.hpp
@@ -11,14 +11,14 @@ struct SPC7110 : Thread {
 
   auto addClocks(uint clocks) -> void;
 
-  auto read(uint24 addr, uint8 data) -> uint8;
-  auto write(uint24 addr, uint8 data) -> void;
+  auto read(uint addr, uint8 data) -> uint8;
+  auto write(uint addr, uint8 data) -> void;
 
-  auto mcuromRead(uint24 addr, uint8 data) -> uint8;
-  auto mcuromWrite(uint24 addr, uint8 data) -> void;
+  auto mcuromRead(uint addr, uint8 data) -> uint8;
+  auto mcuromWrite(uint addr, uint8 data) -> void;
 
-  auto mcuramRead(uint24 addr, uint8 data) -> uint8;
-  auto mcuramWrite(uint24 addr, uint8 data) -> void;
+  auto mcuramRead(uint addr, uint8 data) -> uint8;
+  auto mcuramWrite(uint addr, uint8 data) -> void;
 
   auto serialize(serializer&) -> void;
