+++ bsnes/sfc/coprocessor/st0010/st0010.cpp
@@ -13,11 +13,11 @@ auto ST0010::power() -> void {
   memset(ram, 0x00, sizeof ram);
 }
 
-auto ST0010::read(uint24 addr, uint8 data) -> uint8 {
+auto ST0010::read(uint addr, uint8 data) -> uint8 {
   return readb(addr);
 }
 
-auto ST0010::write(uint24 addr, uint8 data) -> void {
+auto ST0010::write(uint addr, uint8 data) -> void {
   writeb(addr, data);
 
   if((addr & 0xfff) == 0x0021 && (data & 0x80)) {
